AO-377, AO-395: Remove thread local entity cache
There were essentially two options for fixing the broken cache:
 1. Remove it all together and require consumers of ActiveObjects to handle their own caching on top of it
 2. Replace it with a transaction local cache requiring transaction providers to notify the cache when transactions
    start and end

Option 2 will require a lot more thought into how it would be implemented as well as thought into tuning the cache size
and/or making it configurable. This is more complexity than ActiveObjects needs and may not even be an effective cache.
It is a lot easier for consumers to know the shape of their data and how it should be cached.