AO-484 Automatically create prefix indexes on MySQL.
As a premise:
- ActiveObjects does not explicitly specify a row format, which means it
  will always use the default (COMPACT)
- Indexes in MySQL are limited to 767 characters on InnoDB by default
- innodb_large_prefix _does not work_ for COMPACT row format; it only
  works for DYNAMIC and COMPRESSED formats
- By default InnoDB will silently create a prefix index whenever a long
  index is requested, including only the first 255 characters
- Applying innodb_large_prefix causes this silent truncation to become
  an error instead

That means when ActiveObjects successfully creates long indexes, it is
only (and always) because MySQL has silently created a prefix index
instead.

This change codifies that behaviour. Whenever an index is created on a
column which has a string length greater than 255, a prefix index is
explicitly created on the first 255 characters. This ensures the schema
can be created when innodb_large_prefix is enabled.

ON NET, the schema produced with these changes applied is _identiical_
to the schema that has always been produced. No upgrade task is required.